rename and move optional field tests
